Handle enter key press in ShellBrowserImpl
Note that this commit also results in a slight behaviour change when
multiple items are selected and enter is pressed. Previously, all
selected files would be opened externally, while at most a single folder
would be opened in the current tab.

Now, all selected files are still opened externally, but any selected
folders will be opened in new tabs. Note that even if a single folder is
selected (along with one or more files), the folder will still be opened
in a new tab. That's at least consistent - if more than one item is
selected and enter is pressed, any folders will be opened in new tabs
consistently.