Manually link vcpkg .lib dependencies when compiling with Clang
vcpkg's autolink doesn't work with lld-link.exe. This makes it necessary
to manually link the necessary libraries when compiling with Clang.
That's now done here, so that the Clang build now works.

Note that only the x64 build works presently (with warnings). The x86
build fails (seemingly due to
https://github.com/llvm/llvm-project/issues/56507), as does the ARM64
build.

Also note that the dependencies (managed by vcpkg) are still built using
cl. That shouldn't be an issue, though, since clang-cl is designed to be
ABI-compatible with cl.