Add ResourceIconModel class
This class represents a resource icon and can be supplied to MenuView to
display the icon on the menu.

With this commit, the way in which MenuView adds images to the menu has
been updated. Previously, the images would be added immediately.
However, the downside of that was that there was no DPI scaling applied.
That's mostly ok when using ShellIconModel (which loads shell icons),
since the system image list is automatically scaled to match the system
DPI. However, when displaying resource icons, the icon size will need to
be selected based on the DPI.

Since the DPI isn't known until the menu is actually shown, that means
that adding images to the menu has to be deferred until the menu is
shown. Therefore, that's what MenuView now does. When an item is added
to the menu, no image will be set. Just before the menu is actually
shown, the images will be added, with the correct scale.

Ultimately, this means that a client can specify that an image should be
shown on the menu without having any special handling for DPI scaling.