Show help text when using MenuView to display a menu
Previously, there was some code to show help text for submenus nested
within the main menu. However, there was no general solution for showing
help text when using MenuView to display a menu.

Now, when the menu associated with a MenuView instance is shown, the
MenuView instance will register itself as a help text provider. That
will allow MenuView to return the help text for the selected item.