Add ShellBrowserFactory class
An instance of this class is passed to TabContainerImpl. This will make
it easier to instantiate TabContainerImpl in tests, since it will no
longer have a hard dependency on ShellBrowserImpl. Instead,
TabContainerImpl will be able to create ShellBrowser instances via the
provided factory.

This also reduces the number of parameters being passed to the
TabContainerImpl constructor.