Handle copy universal paths command via command target system
This also fixes an issue in which copying a universal path for a
treeview item would always result in the parsing path being copied.
That's because WNetGetUniversalName() ultimately relies on enough space
being allocated by the caller to hold the path that's returned.

When copying a universal path for a treeview item, a small, fixed-size
buffer was previously being used, which was unable to hold anything.
That would then cause the WNetGetUniversalName() call to consistently
fail, meaning that only the parsing path would be copied.