Only start tab drag if mouse moves a certain amount
Previously, the tab drag would be started as soon as the left mouse
button went down. However, that's problematic in certain situations.

For example, if there are more tabs in the control than can be shown on
screen, the last tab is selected and the second last tab is the last
visible tab, when the second last tab is clicked, the tab control will
select it and scroll to the left. That will cause the tab to be
repositioned to the left and a mouse move message will be generated. At
that point, the cursor will be past the right edge of the tab, which
will cause the tab to move to the right. Therefore, simply clicking the
second last tab will result in the tab being moved to the last position.

To fix that, the drag will now only be started if the cursor moves a
certain amount (determined by the SM_CXDRAG and SM_CYDRAG system values)
once the left mouse button goes down. That will prevent a tab from being
moved if the mouse is simply clicked.