Add TreeView class
This class manages a treeview control and is now used with the bookmarks
treeview. This has the significant advantage that the code involved can
be tested.

Note that this also fixes an issue where the bookmarks treeview wouldn't
be updated correctly if a bookmark folder containing nested folders was
pasted. Only the top-level folder would be added to the treeview. That
then meant that if the user attempted to navigate to one of the nested
folders in the listview (shown in the manage bookmarks dialog), the
application would crash, due to a CHECK failure.