Consistently destroy dialog instances in WM_NCDESTROY
Previously, modal dialogs would be declared as stack variables. However,
doing that meant that the class instance associated with a dialog window
would outlive the window.

That was causing problems when adding a bookmark folder via the add
bookmark dialog, for example. Once OK was selected, the dialog would be
closed and the window destroyed. A bookmark folder would then be
created. Because the dialog class instance was still alive, an attempt
was made to add the new folder to the treeview. That would fail (because
the treeview window no longer existed), triggering a CHECK failure and
crash.

To fix issues like that, all dialogs are now destroyed in WM_NCDESTROY.