Fix issue with moving nodes in TreeView
Within TreeView, moving a node is implemented by removing the node, then
re-adding it. This can cause the selection to change (if the selected
node is the node being moved, or one of its descendants). Previously,
that would result in a selection change notification being generated.

That was causing issues within the manage bookmarks dialog.
Specifically, it could result in the following series of events:

- Folder A is being displayed.
- Folder A is then moved into folder B.
- The bookmark treeview handles the move notification, resulting in the
  node for folder A being removed and selection shifting (e.g. to folder
  B).
- As a result of the selection change, the bookmark listview navigates
  to folder B.
- The bookmark listview then handles the original move notification. As
  the item being moved (folder A) is a child of the current folder
  (folder B), it will be added. However, the item already exists, since
  it was added during the navigation in the previous step.

To fix that, the TreeView class now (1) blocks selection change
notifications during a move and (2) restores the original selection once
the moved node has been re-added.