Make ghosted a property of TreeViewNode
Previously, a TreeView node was marked as ghosted by calling a method on
that class. However, it makes more sense for the ghosted value to be a
property of the node itself, not a transient property of the view, since
some items (e.g. hidden files) will be intrinsically ghosted.