Update position of context menu in TreeView
WM_CONTEXTMENU can be sent when the context menu key is pressed. In that
case, the position of the menu will be determined from the bounds of the
selected node.

Previously, the bounds of the node as a whole were used, causing the
context menu to appear at the left edge of the view. Now, the bounds of
the text are used instead. That ensures that the menu is anchored to the
left edge of the text, rather than the left edge of the view.