Update handling of middle-clicks/right-clicks in TreeView
When a middle-click or right-click occurs in TreeView, the click will
now only be handled if the mouse is over a node's icon or text.
Previously, the click would be handled no matter where in a node's row
the mouse was.

Handling these sorts of clicks only when the mouse is over the icon/text
for a node makes more sense, since that matches the visual cues that are
shown (e.g. the node is only highlighted when the mouse is over the
icon/text).