Only set treeview item callback properties after insertion
When a treeview item has a callback property set, the treeview will
dispatch the TVN_GETDISPINFO message to get the property value.

Previously, when inserting an item, the callback properties were set
immediately. However, that could then cause the treeview to dispatch
TVN_GETDISPINFO to get the property values as part of the
TreeView_InsertItem() call. That's problematic, since the mapping from
HTREEITEM to TreeViewNode is only set up once the call to
TreeView_InsertItem() has completed.

Therefore, if the treeview dispatched TVN_GETDISPINFO during the
TreeView_InsertItem() call, a CHECK failure would be triggered, due to
the fact that the HTREEITEM mapping hadn't been set up yet. This was
causing a test failure on AppVeyor.

To fix that, the callback properties will now only be set once the item
has been inserted and the appropriate mapping has been set up.