Sort TreeView items
Rather than inserting items in a specific position, items are now
sorted, with the inserted position determined automatically. This makes
it easier to handle the case where the items need to be sorted in a way
that might differ from their insertion order. This also matches how
items are managed in the ListView.