Store path for PidlAbsolute in debug mode
Pidls are opaque, which makes debugging difficult in general. For
example, if a change notification is received for a particular pidl,
there's no easy way of telling exactly what item the pidl represents.

To improve that, when using PidlAbsolute, the path is now retrieved and
stored in debug mode. Additionally, a small Natvis file has been added
that will customize the way that PidlAbsolute/PidlRelative/PidlChild are
displayed.

When empty, the text "empty" will be displayed. When PidlAbsolute is
non-empty, the path of the pidl will be displayed. When
PidlRelative/PidlChild are non-empty, "(Relative pidl)"/"(Child pidl)"
will be displayed, since there is no way to get a name in those cases.