Add ShellEntry class
This class represents an individual shell item, either a file or a
folder. A folder can contain ShellEntry children, forming a tree.

Additionally, when a folder is loaded, that folder will be watched for
changes. When a change occurs for a particular shell item, the
associated ShellEntry will be updated.

The advantage of this class is that it allows a set of shell items to be
represented as an independent collection, which can then potentially be
used from different places (e.g. from the main treeview and listview).